Pinball Arcade (US Ver1.25) PSN -NoNpDRM?
Still being worked on by myself.
Work perfectly on my PSN Version. Need Testing if they work on NoNpDRM
Dont ask about Ripley's Believe it or Not. For some reason, it goes from Ball 1, then 2, and stays there when used.
Genie and Big/Hot Shot share the same Code. Black Hole also shares the Address, but due to this Table only Set to play 3, I decided to give it a Code of it's own.
For now, all I have left is Elvira and the Party Monsters. This Table for some reason is being hard to find the Code for somehow. It does not Count in Reverse like the last 2 Tables, or follow the rest. Will continue to Search and Test.