Call of Duty-BlackOpsDeclassified (EU) VPK/Mai/NND
Author: Speedfly and Gayfriend
NOTE: Post says.....
Games for the Vitamin format to extract the game and patch patch files covering the contents of the game body and then use the MAI installation
I Tested this on the NoNpDRM Version. Works perfectly. Not sure why, but Speedfly Code Postings put a Space after the $ for the Codes even when you do not have one there.